Crate dlmalloc

A Rust port of the dlmalloc allocator.

The dlmalloc allocator is described at http://g.oswego.edu/dl/html/malloc.html and this Rust crate is a straight port of the C code for the allocator into Rust. The implementation is wrapped up in a Dlmalloc type and has support for Linux, OSX, and Wasm currently.

The primary purpose of this crate is that it serves as the default memory allocator for the wasm32-unknown-unknown target in the standard library. Support for other platforms is largely untested and unused, but is used when testing this crate.

Structs

Dlmalloc

An allocator instance

GlobalDlmalloc

An instance of a “global allocator” backed by Dlmalloc

Traits

Allocator

In order for this crate to efficiently manage memory, it needs a way to communicate with the underlying platform. This Allocator trait provides an interface for this communication.

Functions

enable_alloc_after_fork⚠

allows the allocator to remain unsable in the child process, after a call to fork(2)